refactor: yml code moved to utils.yaml
